refactor: further split classes into files, decoupling
the basic dependency is main/view -> seria_controller/localization -> seria_model
